# file-storage/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для хранения и управления файлами",
        "title": "File Storage API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/download/{id}": {
            "get": {
                "description": "Загружает файл по его уникальному идентификатору в формате multipart/form-data",
                "produces": [
                    "multipart/form-data"
                ],
                "tags": [
                    "Загрузка"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Уникальный идентификатор файла",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.DownloadResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/record": {
            "get": {
                "description": "Получает метаданные всех файлов",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Инфо"
                ],
                "responses": {
                    "200": {
                        "description": "Список метаданных файлов",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.FileRecord"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/record/{id}": {
            "get": {
                "description": "Получает метаданные файла по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Инфо"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Уникальный идентификатор файла",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Метаданные файла",
                        "schema": {
                            "$ref": "#/definitions/models.FileRecord"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/upload/{filename}": {
            "post": {
                "description": "Загружает файл на сервер и возвращает его уникальный идентификатор",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Загрузка"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Имя файла",
                        "name": "filename",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Файл для загрузки",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.UploadResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.DownloadResponse": {
            "type": "object",
            "properties": {
                "record": {
                    "$ref": "#/definitions/models.FileRecord"
                }
            }
        },
        "handlers.UploadResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "models.FileRecord": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}